feat(pricing): add tiered pricing system and normalize reset times
Implement configurable tiered pricing from pricing.json with 2x input
and 1.5x output multipliers above 200K context threshold. Update
pricing.json with Sonnet 4.5, Haiku 4.5, dated variants, and
1-hour cache tiers.

Normalize reset times to nearest :00 when within 1 minute of the hour,
applied across all reset time sources (transcript, cache, OAuth).

+++ src/window.rs
@@ -417,7 +417,7 @@ fn progressive_lookback_block(
         return None;
     }
 
-    let session_duration = Duration::hours(WINDOW_DURATION_HOURS as i64);
+    let session_duration = Duration::hours(WINDOW_DURATION_HOURS);
     let lookback_windows = [
         Duration::hours(10), // 2x session duration - catches most cases
         Duration::hours(24), // Full day for longer sessions
